Best it back into working with dmd 2.061
Most changes are just adding "override" to overridden methods, because
this is deprecated now.

Replace some deprecated Phobos uses.

Not sure if the changes in strparser.d with respect to leading spaces
are actually needed, I didn't pursue it further when it appeared to
just work.

Mask an unknown crash in commandline.d; it's probably a regression and
I can't make sense out of the code.

The worst issue was that calling formattedWrite() in misc.d mysteriously
failed compiling deep inside of Phobos. For some reason it works when
providing something that can handle put() with both char[] and dchar.
I would claim that normally, an output range for just the type "char"
should work (as the format string is immutable(char)[]). I guess this
ranges crap is really fragile, and I'm glad that I added wrappers for
it, so that only misc.d has to be changed.

The second worst issue was isIntegral() returning true for enum types.
I really wonder whether this is intentional, or whether it's a new
Phobos bug. isBasicType() returns true for enums too. Just work it
around by using the "awesome" is(T==enum) expression. This issue lead
to weird failures of Lua-to-D invocations, because enums are handled
completely differently from integers.

There are some order of declaration issues in color.d. This used to
work fine even in old D1 dmd versions.

In time.d the signature of opEquals had to be changed. It seems earlier
versions of dmd could convert lvalues to "const ref", while the dmd
2.061 can't. (I don't really see any reason why it can't...)

This compiles without warnings, and seems to run fine. Only a few
things have been actually tested, and it's very possible that there are
many bugs left.

All in all, I'm very disappointed by D(2), and how the language and
compiler perform in this supposedly stable state. It seems const()
and the templatization of Phobos had mainly bad consequences. The
effort for converting this project to D2 could have better spent by
targetting another language instead, even if that would have been more
work.